+++ root/recording/edit_form.tt
@@ -12,7 +12,7 @@
 </div>
 
 [% script_manifest('edit.js') %]
-[% script_manifest('recording/form.js', {async => 'async'}) %]
+[% script_manifest('recording/edit.js', {async => 'async'}) %]
 
 <p>[%- l('For more information, check the {doc_doc|documentation} and {doc_styleguide|style guidelines}.', {doc_doc => doc_link('Recording'), doc_styleguide => doc_link('Style/Recording')}) -%]</p>
