+++ root/static/scripts/recording/edit.js
