+++ webpack/client.config.js
@@ -36,40 +36,40 @@ const entries = [
   'account/applications/register',
   'account/edit',
   'account/preferences',
-  'area',
   'alias',
+  'area/edit',
   'area/index',
   'area/places-map',
-  'artist',
+  'artist/edit',
   'artist/index',
   'collection/edit',
   'common',
   'confirm-seed',
   'edit',
   'edit/notes-received',
-  'event',
+  'event/edit',
   'event/index',
   'instrument/index',
   'jed-data',
   'label/index',
-  'place',
+  'place/edit',
   'place/index',
   'place/map',
-  'recording/form',
+  'recording/edit',
   'recording/merge',
   'register',
-  'release/coverart',
-  'release/index',
   'release-editor',
   'release-group/index',
-  'series',
+  'release/coverart',
+  'release/index',
+  'series/edit',
   'series/index',
   'statistics',
   'timeline',
-  'url',
+  'url/edit',
   'user/login',
   'voting',
-  'work',
+  'work/edit',
   'work/index',
 ].reduce((accum, name) => {
   accum[name] = path.resolve(dirs.SCRIPTS, `${name}.js`);
